Added: Technician List, Delete Actions

+++ frontend/src/components/Dashboard/MainContent.jsx
@@ -2,6 +2,10 @@ import React, { useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 
 import { deleteUser, listUsers } from "../../redux/actions/userActions";
+import {
+  technicianDelete,
+  listTechnicians,
+} from "../../redux/actions/technicianActions";
 
 import Loader from "../../components/Loader";
 import Message from "../../components/Message.jsx";
@@ -9,9 +13,9 @@ import Table from "../Table";
 
 const MainContent = ({ variant, selectedItem, menuItems }) => {
   const userColumns = ["name", "_id", "email"];
-  const adminColumns = ["name", "_id", "email"];
+  // const adminColumns = ["name", "_id", "email"];
   const technicianColumns = ["name", "_id", "email"];
-  const orderColumns = ["orderId", "customerId", "status"];
+  // const orderColumns = ["orderId", "customerId", "status"];
 
   const dispatch = useDispatch();
 
@@ -51,15 +55,20 @@ const MainContent = ({ variant, selectedItem, menuItems }) => {
     technicians,
   } = technicianUserList;
 
-  // Delete user
   const handleDeleteUser = (userId) => {
-    dispatch(deleteUser(userId)).then(() => {
-      dispatch(listUsers());
-    });
+    if (window.confirm("Are You Sure?")) {
+      dispatch(deleteUser(userId)).then(() => {
+        dispatch(listUsers(adminUserInfo.user._id));
+      });
+    }
   };
 
   const handleDeleteTechnician = (technicianId) => {
-    console.log("Delete Technician");
+    if (window.confirm("Are You Sure?")) {
+      dispatch(technicianDelete(technicianId)).then(() => {
+        dispatch(listTechnicians());
+      });
+    }
   };
 
   return (
@@ -80,14 +89,9 @@ const MainContent = ({ variant, selectedItem, menuItems }) => {
         </h2>
       )}
 
-      {usersLoading || technicianUsersLoading ? (
-        <Loader />
-      ) : usersError || technicianUsersError ? (
-        <Message variant="error">{usersError}</Message>
-      ) : (
-        selectedItem !== null && (
-          <div>
-            {/* {menuItems[selectedItem].name === "Admins" && (
+      {selectedItem !== null && (
+        <div>
+          {/* {menuItems[selectedItem].name === "Admins" && (
               <div>
                 <h2 className="text-2xl font-semibold mb-2">Admins</h2>
                 {admins.map((admin) => (
@@ -95,70 +99,85 @@ const MainContent = ({ variant, selectedItem, menuItems }) => {
                 ))}
               </div>
             )} */}
-            {menuItems[selectedItem].name === "Technicians" && (
-              <div>
-                <h2 className="text-2xl font-semibold mb-2">All Technicians</h2>
+          {menuItems[selectedItem].name === "Technicians" && technicians && (
+            <div>
+              <h2 className="text-2xl font-semibold mb-2">All Technicians</h2>
+              {technicianUsersLoading ? (
+                <Loader />
+              ) : technicianUsersError ? (
+                <Message variant="error">{technicianUsersError}</Message>
+              ) : (
                 <Table
                   data={technicians}
                   columns={technicianColumns}
                   handleDelete={handleDeleteTechnician}
                   entityType="technician"
                 />
-              </div>
-            )}
-            {menuItems[selectedItem].name === "Users" && users && (
-              <div>
-                <h2 className="text-2xl font-semibold mb-2">All Users</h2>
+              )}
+            </div>
+          )}
+
+          {menuItems[selectedItem].name === "Users" && users && (
+            <div>
+              <h2 className="text-2xl font-semibold mb-2">All Users</h2>
+              {usersLoading ? (
+                <Loader />
+              ) : usersError ? (
+                <Message variant="error">{usersError}</Message>
+              ) : (
                 <Table
                   data={users}
                   columns={userColumns}
                   handleDelete={handleDeleteUser}
                   entityType="user"
                 />
-              </div>
-            )}
-            {menuItems[selectedItem].name === "Orders" && (
-              <div>
-                <h2 className="text-2xl font-semibold mb-2">Orders</h2>
-                {variant === "admin" && (
-                  <div>
-                    <h3 className="text-xl font-semibold mb-2">All Orders</h3>
-                    {allOrders.map((order) => (
-                      <div key={order.id}>{order.name}</div>
-                    ))}
-                  </div>
-                )}
-                {variant === "user" && (
-                  <div>
-                    <h3 className="text-xl font-semibold mb-2">Your Orders</h3>
-                    {userOrders.map((order) => (
-                      <div key={order.id}>{order.name}</div>
-                    ))}
-                  </div>
-                )}
-                {variant === "technician" && (
-                  <div>
-                    <h3 className="text-xl font-semibold mb-2">
-                      Your Assigned Orders
-                    </h3>
-                    {technicianOrders.map((order) => (
-                      <div key={order.id}>{order.name}</div>
-                    ))}
-                  </div>
-                )}
-              </div>
-            )}
+              )}
+            </div>
+          )}
 
-            {menuItems[selectedItem].name === "Reviews" && (
-              <div>
-                <h2 className="text-2xl font-semibold mb-2">Users</h2>
-                {users.map((user) => (
-                  <div key={user.id}>{user.name}</div>
-                ))}
-              </div>
-            )}
-          </div>
-        )
+          {menuItems[selectedItem].name === "Orders" && (
+            <div>
+              <h2 className="text-2xl font-semibold mb-2">Orders</h2>
+              {variant === "admin" && (
+                <div>
+                  <h3 className="text-xl font-semibold mb-2">All Orders</h3>
+                  {allOrders.map((order) => (
+                    <div key={order.id}>{order.name}</div>
+                  ))}
+                </div>
+              )}
+
+              {variant === "user" && (
+                <div>
+                  <h3 className="text-xl font-semibold mb-2">Your Orders</h3>
+                  {userOrders.map((order) => (
+                    <div key={order.id}>{order.name}</div>
+                  ))}
+                </div>
+              )}
+
+              {variant === "technician" && (
+                <div>
+                  <h3 className="text-xl font-semibold mb-2">
+                    Your Assigned Orders
+                  </h3>
+                  {technicianOrders.map((order) => (
+                    <div key={order.id}>{order.name}</div>
+                  ))}
+                </div>
+              )}
+            </div>
+          )}
+
+          {menuItems[selectedItem].name === "Reviews" && (
+            <div>
+              <h2 className="text-2xl font-semibold mb-2">Users</h2>
+              {users.map((user) => (
+                <div key={user.id}>{user.name}</div>
+              ))}
+            </div>
+          )}
+        </div>
       )}
     </div>
   );
